added default value handling
