+++ index.html
@@ -7,7 +7,6 @@
 <body>
 <nav>
 <input id=jsOpen type="file" accept=".js">
-<input id=picoOpen type="file" accept=".js">
 <input id=fileOpen type="file" accept=".json">
 <button id=fileSave >Save</button>
 </nav>
